Phases 6 and 7: RESP server and YCSB benchmark
Phase 6 turns the embedded store into a server. `lsmrs serve [addr]` speaks a
RESP2 subset -- PING, ECHO, GET, SET, DEL, QUIT, COMMAND -- so redis-cli and
other Redis clients drive it unmodified.

Db::get already took &self and put/delete took &mut self, so Arc<RwLock<Db>>
maps onto the existing API with no changes: reads run concurrently, writes take
the exclusive lock. Thread per connection, with try_clone splitting the socket
so the reader and writer halves can be owned independently.

resp.rs parses RESP arrays and inline commands. A bulk-string length is an
allocation instruction from an untrusted party, so MAX_BULK_LEN is checked
before the allocation, not after -- the first untrusted input in this project.
Length prefixes also make the protocol binary-safe: a value containing \r\n,
the protocol's own terminator, round-trips intact.

Command errors reply and keep the connection; protocol errors reply and hang
up, because a desynchronised stream cannot be resynchronised. Shutdown wakes
the blocked accept by connecting to its own port, which also makes the server
joinable in tests rather than leaking a thread each.

Phase 7 adds a YCSB-style load generator that speaks RESP, so the same binary
drives lsmrs or a real redis-server: workloads A/B/C, Zipfian or uniform keys
scrambled through the store's own hash, per-operation latency percentiles.

The numbers, 20k records / 40k ops / 8 threads / Zipfian, 16 cores:

  lsmrs, fsync per write        4,011 ops/sec   read p50 2031us
  redis, appendfsync always     9,427 ops/sec   read p50  876us
  lsmrs, no fsync             172,971 ops/sec   read p50   30us
  redis, no AOF               123,934 ops/sec   read p50   60us

fsync is worth 43x, which dwarfs every in-process optimisation in the project.
Redis beats lsmrs 2.3x at equal durability while single-threaded, because it
group-commits its AOF: N concurrent writers share one disk round-trip, where
lsmrs fsyncs inside put under the global write lock and pays N. That is the
largest piece of known work left. Where lsmrs is ahead it is 8 threads against
one on a 16-core box, which is arithmetic rather than engineering -- and Redis
still has the tighter tail, having neither lock contention nor a compactor.

Also fixes the last clippy warning in wal.rs, brings README and DESIGN.md up to
date through all seven phases, and records the full analysis in NOTES.md.

+++ src/wal.rs
@@ -138,7 +138,7 @@ impl Wal {
             }
             let length = u32::from_le_bytes(length_bytes);
             let mut checksum_bytes = [0u8; 4];
-            if length < CRC_SIZE as u32 {
+            if length < CRC_SIZE {
                 break;
             }
             match reader.read_exact(&mut checksum_bytes) {
